Don't fail on workflow 404
We have an issue where we have a listed 'running' workflow which has
expired and we can no longer look it up. This is a bug in the CircleCI
API.

To work around this while we report it to them, this change should
consider 404s valid responses for a workflow lookup, and just not take
them into account when queueing.